requirements.txt
ovos_lingua_franca